Added a Few Fixes After Class
The Remove button for existing invoice items was not showing up.  Fix : Add Editable="false" to the TemplateColumn tag

It was not possible to adjust the Quantity of the items on the invoice to more than 0.  This was a problem with the way the max was being calculated, referencing the parts collection.  Fix : Added a second collection that would actually be adjusted so that the parts items would never be deleted and thus not referenced for the Max value when moving the part into the invoice.  A few changes elsewhere were required to make it all function as intended, including adding an initial set of parts to the parts collection to account for parts that already exist on the invoice so the quantities can be changed for those as well, else they will go to zero immediately on change.

Saving was not removing items from the Invoice display because they were being removed from the invoice object before saving instead of just setting the RemoveFromViewFlag to true.  This has been changed and a few other references to the RemoveFromViewFlag have been added to ensure consistency with a few functions of the InvoiceEdit page.